requirements.txt
Django==1.7.7
psycopg2==2.6
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.3.0
static3==0.5.1
wsgiref==0.1.2
djangorestframework==3.1.1
django-extensions==1.5.2
